Abstracted Vulkan Logic.
Moved all Vulkan logic to its own class. This is now an unique pointer in the engine class.
